+++ python/python/glide/async_commands/core.py
@@ -65,7 +65,7 @@ from glide.constants import (
 from glide.protobuf.command_request_pb2 import RequestType
 from glide.routes import Route
 
-from ..glide import ClusterScanCursor, Script
+from ..glide import ClusterScanCursor
 
 
 class ConditionalChange(Enum):
@@ -5392,36 +5392,6 @@ class CoreCommands(Protocol):
             int,
             await self._execute_command(RequestType.ZInterCard, args),
         )
-
-    async def invoke_script(
-        self,
-        script: Script,
-        keys: Optional[List[TEncodable]] = None,
-        args: Optional[List[TEncodable]] = None,
-    ) -> TResult:
-        """
-        Invokes a Lua script with its keys and arguments.
-        This method simplifies the process of invoking scripts on a the server by using an object that represents a Lua script.
-        The script loading, argument preparation, and execution will all be handled internally.
-        If the script has not already been loaded, it will be loaded automatically using the `SCRIPT LOAD` command.
-        After that, it will be invoked using the `EVALSHA` command.
-
-        See https://valkey.io/commands/script-load/ and https://valkey.io/commands/evalsha/ for more details.
-
-        Args:
-            script (Script): The Lua script to execute.
-            keys (Optional[List[TEncodable]]): The keys that are used in the script.
-            args (Optional[List[TEncodable]]): The arguments for the script.
-
-        Returns:
-            TResult: a value that depends on the script that was executed.
-
-        Examples:
-            >>> lua_script = Script("return { KEYS[1], ARGV[1] }")
-            >>> await invoke_script(lua_script, keys=["foo"], args=["bar"] );
-                [b"foo", b"bar"]
-        """
-        return await self._execute_script(script.get_hash(), keys, args)
 
     async def pfadd(self, key: TEncodable, elements: List[TEncodable]) -> int:
         """
